Add code to trap exit signals from child processes
Previously, when we killed the `HttpServer` process, we **also** kill
the `KickStarter` process because we had linked the two processes.

By trapping the exit signal, we now keep the `KickStarter` process
alive, but we also observe an error message about an unrecognized
message being received by the `handle_info` function.

We will address this issue now.